Move functionality to its own component
