+++ src/components/TheUserMenu.vue
@@ -18,29 +18,8 @@
         {{ authStore.username }}
       </v-btn>
     </template>
+    <LanguageSelection />
     <v-list density="compact" class="bg-surface-2 text-surface-1">
-      <v-menu location="end">
-        <template #activator="{ props }">
-          <v-list-item
-            data-cy="language-menu"
-            v-bind="props"
-            :prepend-icon="`mdi-web`"
-            append-icon="mdi-chevron-right"
-            :title="$i18n.locale"
-          />
-        </template>
-
-        <v-list data-cy="lang-list" density="compact" class="bg-surface-2 text-surface-1">
-          <v-list-item
-            v-for="(lang, i) in $i18n.availableLocales"
-            :key="`${i}-${lang}`"
-            :value="lang"
-            :title="lang"
-            :data-lang="lang"
-            @click="changeLocale(lang)"
-          />
-        </v-list>
-      </v-menu>
       <v-list-item
         v-for="({ text, icon, page, cyName }, i) in menuItems"
         :key="i"
@@ -59,10 +38,12 @@ import { useAuthStore } from '@/stores/auth';
 import { computed } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { ROUTE_NAME_LOGOUT } from '@/router.js';
-import { setLocalStorage } from '../../utils/local-storage-utils';
 
+import LanguageSelection from './LanguageSelection.vue';
 
 const authStore = useAuthStore();
+
+const { t } = useI18n();
 
 defineProps({
   variant:{
@@ -70,15 +51,6 @@ defineProps({
       default:'light'
     }
 });
-
-
-const { t, locale } = useI18n();
-
-const changeLocale = (lang) => {
-  locale.value = lang;
-  setLocalStorage('preferredLocale', lang);
-};
-
 
 const menuItems = computed(() => {
   return [{ text: t('global.logout'), icon: 'logout', page: { name: ROUTE_NAME_LOGOUT }, cyName: 'Log Out' }];
